+++ lib/classic_country_picker.dart
@@ -2,3 +2,5 @@ export 'src/models/country.dart';
 export 'src/data/country_data.dart';
 export 'src/widgets/classic_country_dropdown.dart';
 export 'src/widgets/country_picker_dialog.dart';
+export 'src/widgets/show_country_picker.dart';
+export 'src/utils/country_parser.dart';